Updated webcam library with optimized routines
The webcam library will re-use the given application frame's buffer,
saving us a lot of calloc/free's.

Updated Makefile to add gdb symbols for debugging.
One issue though, is that it consumes 180% CPU while running. Tried to
add a SDL_Delay which helped things a bit, but not by that much.